clean code

+++ src/main/java/cn/nukkit/network/protocol/SetScoreboardIdentityPacket.java
@@ -47,7 +47,7 @@ public class SetScoreboardIdentityPacket extends DataPacket {
     
     public static class SetScoreboardIdentityEntry {
 
-        public int scoreboardId;
+        public long scoreboardId;
         public long entityUniqueId;
     }
 }
